adjusted view all items
