+++ collectMe-UI/src/components/Category.js
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 function Category ({ category }) {
     return (
         <div className="d-flex jusitfy-content-end">
-            {category.categoryName}
+            <Link to="/category/{category.categoryId}">{category.categoryName}</Link>
         </div>
     );
 }
